Use uint32 for console coordinates and fix dimension bug

+++ src/gopheros/device/video/console/device.go
@@ -15,7 +15,7 @@ const (
 // consoles.
 type Device interface {
 	// Dimensions returns the width and height of the console in characters.
-	Dimensions() (uint16, uint16)
+	Dimensions() (uint32, uint32)
 
 	// DefaultColors returns the default foreground and background colors
 	// used by this console.
@@ -24,16 +24,16 @@ type Device interface {
 	// Fill sets the contents of the specified rectangular region to the
 	// requested color. Both x and y coordinates are 1-based (top-left
 	// corner has coordinates 1,1).
-	Fill(x, y, width, height uint16, fg, bg uint8)
+	Fill(x, y, width, height uint32, fg, bg uint8)
 
 	// Scroll the console contents to the specified direction. The caller
 	// is responsible for updating (e.g. clear or replace) the contents of
 	// the region that was scrolled.
-	Scroll(dir ScrollDir, lines uint16)
+	Scroll(dir ScrollDir, lines uint32)
 
 	// Write a char to the specified location. Both x and y coordinates are
 	// 1-based (top-left corner has coordinates 1,1).
-	Write(ch byte, fg, bg uint8, x, y uint16)
+	Write(ch byte, fg, bg uint8, x, y uint32)
 
 	// Palette returns the active color palette for this console.
 	Palette() color.Palette
